Added proper COM port disconnection on exit - MainWindow.xaml.cs
Made changes to the app's exit logic. Now, when you exit the app, it sends a DTR signal to the Arduino, resetting it and properly disconnecting from the COM port. This helps the Arduino controller know when the app will connect next time. Without this change, the controller doesn't know when to send the initial serial data when the app starts.